update utils.py and ca.py for Py3 only, drop utils2 and utils3
