+++ epics/ca.py
@@ -26,10 +26,10 @@ import warnings
 
 from math import log10
 from pkg_resources import resource_filename
+from copy import deepcopy
 
-from .utils import (STR2BYTES, BYTES2STR, NULLCHAR_2,
-                    strjoin, memcopy, is_string, is_string_or_bytes,
-                    ascii_string, clib_search_path)
+from .utils import (str2bytes, bytese2str, strjoin, is_string,
+                    is_string_or_bytes, ascii_string, clib_search_path)
 
 # ignore warning about item size... for now??
 warnings.filterwarnings('ignore',
@@ -469,7 +469,7 @@ def _get_cache_by_chid(chid):
         # context cache before giving up. This branch should not happen often.
         context = current_context()
         if context is not None:
-            pvname = BYTES2STR(libca.ca_name(dbr.chid_t(chid)))
+            pvname = bytes2str(libca.ca_name(dbr.chid_t(chid)))
             return _cache[context][pvname]
         raise
 
@@ -744,7 +744,7 @@ def _onGetEvent(args, **kws):
     elif dbr.IRON_PYTHON:
         result = dbr.cast_args(args)
     else:
-        result = memcopy(dbr.cast_args(args))
+        result = deepcopy(dbr.cast_args(args))
 
     with entry.lock:
         entry.get_results[ftype][0] = result
@@ -885,13 +885,13 @@ def flush_io():
 def message(status):
     """Print a message corresponding to a Channel Access status return value.
     """
-    return BYTES2STR(libca.ca_message(status))
+    return bytes2str(libca.ca_message(status))
 
 @withCA
 def version():
     """   Print Channel Access version string.
     Currently, this should report '4.13' """
-    return BYTES2STR(libca.ca_version())
+    return bytes2str(libca.ca_version())
 
 @withCA
 def pend_io(timeout=1.0):
@@ -997,7 +997,7 @@ def create_channel(pvname, connect=False, auto_cb=True, callback=None):
             chid = dbr.chid_t()
             with entry.lock:
                 ret = libca.ca_create_channel(
-                    ctypes.c_char_p(STR2BYTES(pvname)), _CB_CONNECT, 0, 0,
+                    ctypes.c_char_p(str2bytes(pvname)), _CB_CONNECT, 0, 0,
                     ctypes.byref(chid)
                 )
                 PySEVCHK('create_channel', ret)
@@ -1107,12 +1107,12 @@ def _chid_to_int(chid):
 @withCHID
 def name(chid):
     "return PV name for channel name"
-    return BYTES2STR(libca.ca_name(chid))
+    return bytes2str(libca.ca_name(chid))
 
 @withCHID
 def host_name(chid):
     "return host name and port serving Channel"
-    return BYTES2STR(libca.ca_host_name(chid))
+    return bytes2str(libca.ca_host_name(chid))
 
 @withCHID
 def element_count(chid):
@@ -1214,9 +1214,9 @@ def _unpack(chid, data, count=None, ftype=None, as_numpy=True):
         """ Scan a string, or an array of strings as a list, depending on content """
         out = []
         for elem in range(min(count, len(data))):
-            this = strjoin('', BYTES2STR(data[elem].value)).rstrip()
-            if NULLCHAR_2 in this:
-                this = this[:this.index(NULLCHAR_2)]
+            this = strjoin('', bytes2str(data[elem].value)).rstrip()
+            if '\x00' in this:
+                this = this[:this.index('\x00')]
             out.append(this)
         if len(out) == 1:
             out = out[0]
@@ -1230,9 +1230,9 @@ def _unpack(chid, data, count=None, ftype=None, as_numpy=True):
                 out = numpy.empty(shape=(count,), dtype=dbr.NP_Map[ntype])
                 ctypes.memmove(out.ctypes.data, data, out.nbytes)
             else:
-                out = numpy.ctypeslib.as_array(memcopy(data))
+                out = numpy.ctypeslib.as_array(deepcopy(data))
         else:
-            out = memcopy(data)
+            out = deepcopy(data)
         return out
 
     def unpack(data, count, ntype, use_numpy, elem_count):
@@ -1302,10 +1302,10 @@ def _unpack_metadata(ftype, dbr_value):
             if hasattr(dbr_value, attr):
                 md[attr] = getattr(dbr_value, attr)
                 if attr == 'units':
-                    md[attr] = BYTES2STR(getattr(dbr_value, attr, None))
+                    md[attr] = bytes2str(getattr(dbr_value, attr, None))
 
         if hasattr(dbr_value, 'strs') and getattr(dbr_value, 'no_str', 0) > 0:
-            md['enum_strs'] = tuple(BYTES2STR(dbr_value.strs[i].value)
+            md['enum_strs'] = tuple(bytes2str(dbr_value.strs[i].value)
                                     for i in range(dbr_value.no_str))
     elif ftype >= dbr.TIME_STRING:
         md['status'] = dbr_value.status
@@ -1543,7 +1543,7 @@ def get_complete_with_metadata(chid, ftype=None, count=None, timeout=None,
     if as_string:
         val = _as_string(val, chid, count, ftype)
     elif isinstance(val, ctypes.Array) and HAS_NUMPY and as_numpy:
-        val = numpy.ctypeslib.as_array(memcopy(val))
+        val = numpy.ctypeslib.as_array(deepcopy(val))
 
     # value retrieved, clear cached value
     metadata['value'] = val
@@ -2000,7 +2000,7 @@ def sg_put(gid, chid, value):
     else:
         # auto-convert strings to arrays for character waveforms
         # could consider using
-        # numpy.fromstring(("%s%s" % (s,NULLCHAR*maxlen))[:maxlen],
+        # numpy.fromstring(("%s%s" % (s, pythonb'\x00'*maxlen))[:maxlen],
         #                  dtype=numpy.uint8)
         if ftype == dbr.CHAR and is_string_or_bytes(value):
             pad = [0]*(1+count-len(value))
